refactor(Lotto): Lotto 클래스를 기존 제공 받은 경로에 위치하도록 변경
