+++ src/domain/index.js
@@ -3,4 +3,3 @@ export * from './confirmWinningResult/index.js';
 
 // file
 export { default as lottoPurchase } from './lottoPurchase/lottoPurchase.module.js';
-export { default as Lotto } from './lotto.module.js';
