+++ src/validations/index.js
@@ -1,4 +1,4 @@
 export { default as bonusNumberValidation } from './bonusNumberValidation/bonusNumberValidation.module.js';
 export { default as commonValidation } from './commonValidation/commonValidation.module.js';
-export { default as lottoNumberValidation } from './lottoNumberValidation/lottoNumberValidation.module.js';
 export { default as purchasedLottoAmountValidation } from './purchasedLottoAmountValidation/purchasedLottoAmountValidation.module.js';
+export { default as lottoNumberValidation } from './lottoNumberValidation/lottoNumberValidation.module.js';
